Time: 155 ms (40.28%), Space: 63.1 MB (37.91%) - LeetHub

+++ 2616-maximal-score-after-applying-k-operations/2616-maximal-score-after-applying-k-operations.java
@@ -1,16 +1,16 @@
 class Solution {
     public long maxKelements(int[] nums, int k) {
-        PriorityQueue<Long> maxHeap = new PriorityQueue<>(Collections.reverseOrder());
+        PriorityQueue<Integer> maxHeap = new PriorityQueue<>(Collections.reverseOrder());
         for(int i = 0 ; i<nums.length ;i++){
-            maxHeap.add((long)nums[i]) ; 
+            maxHeap.add(nums[i]) ; 
         }
         long score = 0 ;
         while(k>0){
-            long x = maxHeap.poll() ; 
+            int x = maxHeap.poll() ; 
             score +=x ;
            // System.out.println(x);
           // System.out.println(Math.ceil(x/3));
-            maxHeap.add((long)Math.ceil(x/3.0)) ;
+            maxHeap.add((int)Math.ceil(x/3.0)) ;
             k-- ;
         }
         return score;
